rename RecordCardsContainer > SubTypeAccordionContainer
